# public/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vertifile API",
    "version": "4.1.0",
    "description": "Document protection and verification platform. Vertifile creates tamper-proof PVF (Protected Verification File) documents with HMAC-SHA256 signatures, blind hashing, and optional blockchain anchoring on Polygon.",
    "contact": {
      "name": "Zur Halfon",
      "email": "zur2525@gmail.com",
      "url": "https://vertifile.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://vertifile.com",
      "description": "Production"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
    "/api/signup": {
      "post": {
        "tags": ["System"],
        "summary": "Sign up for an API key",
        "description": "Register a new organization and receive an API key. Pro plan allows 100 requests per day. Limited to 3 signups per IP per day.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signup successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/verify": {
      "post": {
        "tags": ["Documents"],
        "summary": "Verify a document hash",
        "description": "Public endpoint. Verify a document by its SHA-256 hash. Optionally include the HMAC signature, recipient hash, and code integrity token for full verification. Returns a fresh session token on success. Rate limited to 500 requests per 15 minutes.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key obtained from POST /api/signup. Prefix: vf_live_"
      }
    },
    "schemas": {
      "SignupRequest": {
        "type": "object",
        "required": ["orgName", "contactName", "email"],
        "properties": {
          "orgName": {
            "type": "string",
            "maxLength": 100,
            "description": "Organization name",
            "example": "Acme University"
          },
          "contactName": {
            "type": "string",
            "maxLength": 100,
            "description": "Contact person name",
            "example": "Jane Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Contact email address",
            "example": "jane@acme-university.edu"
          },
          "useCase": {
            "type": "string",
            "description": "Brief description of intended use",
            "example": "Issuing verifiable academic transcripts"
          }
        }
      },
      "SignupResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "apiKey": {
            "type": "string",
            "description": "API key (shown only once)",
            "example": "vf_live_a1b2c3d4e5f6..."
          },
          "orgId": {
            "type": "string",
            "description": "Unique organization identifier",
            "example": "org_acme_university_1a2b3c4d"
          },
          "orgName": { "type": "string", "example": "Acme University" },
          "plan": {
            "type": "string",
            "enum": ["pro", "business", "enterprise"],
            "example": "pro"
          },
          "rateLimit": {
            "type": "integer",
            "description": "Requests per rate-limit window",
            "example": 5
          },
          "message": { "type": "string", "example": "Save this API key — it will not be shown again." }
        }
      },
...
      "VerifyRequest": {
        "type": "object",
        "properties": {
          "hash": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "description": "SHA-256 hash of the document content",
            "example": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
          },
          "signature": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "description": "HMAC-SHA256 signature"
          },
          "content": {
            "type": "object",
            "description": "Legacy: content object (hash will be computed server-side)"
          },
          "recipientHash": {
            "type": "string",
            "description": "SHA-256 hash of the recipient email for binding verification"
          },
          "codeIntegrity": {
            "type": "string",
            "description": "Code integrity token from the PVF viewer"
          },
          "created": {
            "type": "string",
            "description": "Creation timestamp"
          },
          "orgId": {
            "type": "string",
            "description": "Organization ID of the issuer"
          }
        }
      },
      "VerifyResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "verified": {
            "type": "boolean",
            "description": "Whether the document passed verification"
          },
          "hash": {
            "type": "string",
            "description": "The hash that was looked up"
          },
          "reason": {
            "type": "string",
            "enum": ["invalid_signature", "code_tampered", "chain_broken", "recipient_mismatch"],
            "description": "Failure reason (only present when verified is false)"
          },
          "token": {
            "type": "string",
            "description": "Fresh session token (only present when verified is true)"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Document creation timestamp"
          },
          "orgName": {
            "type": "string",
            "description": "Issuing organization name"
          },
          "blockchain": {
            "type": "object",
            "nullable": true,
            "description": "Blockchain proof (if on-chain mode is active)",
            "properties": {
              "verified": { "type": "boolean" },
              "txHash": { "type": "string" },
              "network": { "type": "string" }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "description": "Human-readable error message" }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request parameters",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "error": "Missing required field" }
          }
        }
      },
...
      "RateLimited": {
        "description": "Rate limit exceeded",
        "headers": {
          "RateLimit-Limit": {
            "description": "Maximum requests allowed in the window",
            "schema": { "type": "integer" }
          },
          "RateLimit-Remaining": {
            "description": "Remaining requests in the current window",
            "schema": { "type": "integer" }
          },
          "RateLimit-Reset": {
            "description": "Seconds until the rate limit resets",
            "schema": { "type": "integer" }
          }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "error": "Too many requests, try again later" }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "error": "Internal server error" }
          }
        }
      }
    }
  }
}